[#175] Renamed various files to make their function clearer
